+++ genotype_api/api/endpoints/plates.py
@@ -17,10 +17,10 @@ from genotype_api.database.crud.read import (
 )
 from genotype_api.database.crud.update import refresh_sample_status
 from genotype_api.database.crud.create import create_plate, create_analyses_sample_objects
-from genotype_api.database.database import get_session
+from genotype_api.database.session_handler import get_session
 from genotype_api.file_parsing.excel import GenotypeAnalysis
 from genotype_api.file_parsing.files import check_file
-from genotype_api.database.models.models import (
+from genotype_api.database.models import (
     Analysis,
     User,
     Plate,
