+++ genotype_api/file_parsing/excel.py
@@ -6,7 +6,7 @@ from typing import ByteString, Iterable, List, Optional
 
 import openpyxl
 from genotype_api.exceptions import SexConflictError
-from genotype_api.database.models.models import Genotype, Analysis
+from genotype_api.database.models import Genotype, Analysis
 
 from openpyxl.workbook import Workbook
 from openpyxl.worksheet.worksheet import Worksheet
