+++ genotype_api/security.py
@@ -7,8 +7,8 @@ from starlette import status
 
 from starlette.requests import Request
 
-from genotype_api.database.database import get_session
-from genotype_api.database.models.models import User
+from genotype_api.database.session_handler import get_session
+from genotype_api.database.models import User
 from genotype_api.config import security_settings
 from genotype_api.database.crud.read import get_user_by_email
 from jose import jwt
